+++ NEWS.md
@@ -1,5 +1,8 @@
 # devtools 1.12.0.9000
 
+* New `use_gpl3_license()` sets the license field in `DESCRIPTION` and
+  includes a copy of the license in `LICENSE`.
+
 * Improve Git status checks used in `release()` (#1205, @krlmlr).
 
 * Various minor improvements around checking of reverse dependencies